Fixed the way id3-go writes back comm tags when using utf16
Previously when trying to write a null terminated string it simply wrote
a single 0 byte for the null character. This fails for utf16 where the
null character is two bytes of zeros.

fixed the way that we WriteNullTermString with utf16

Originally this function was just writing a 0x00 byte, which does not
work for utf16.

fixed mistake when merging changes