cleaned up testing of UTF-16 on UnsynchTextFrames fix
Added a new COMM field to test.mp3 in UTF-16 with the Description: "✓"
and text "✓". Opening and closing it two times without a panic should
ensure that it is working.

Also added a few lines to the open test to explicitly test to make sure
that these UTF-16 fields are being read properly.